fix: PoseStack crash on some occasions (#10)
* fix: Use long instead of int for numeric fields in GuildRaid

* chore: Bump mod version

* feat: Add hover tooltips for raid features in the config

* chore: Send mod version as a number in GIdentifyWSMessage

* refactor: Send online guild members as one message, so it can be copied easier

* fix: Fix incorrect lang message

* fix: Pose stack fail safe

* refactor: Improve UUID parsing error logging for MojangService

* refactor: Simplify regex patterns for Buster connection messages

* fix: Shadowling regex for TNARaidFeature

+++ common/src/main/java/dev/lotnest/sequoia/feature/features/guildraidtracker/GuildRaid.java
@@ -8,4 +8,4 @@ import java.util.List;
 import java.util.UUID;
 
 public record GuildRaid(
-        RaidType type, List<String> players, UUID reporterID, int aspects, int emeralds, int xp, int sr) {}
+        RaidType type, List<String> players, UUID reporterID, long aspects, long emeralds, long xp, long sr) {}
